Adding questions about expressions
I just go tto thinking that it is a very different think to have a Sexual Orientation or a Gender Identity that is different to the majority and then to take that further and do something about it. It will be interesting to see if there is any difference in the experiences we get back from this.

(can you tell I'm trying really hard not to say "in or out of the closet"?)